+++ app/policies/offer_policy.rb
@@ -1,6 +1,12 @@
 # frozen_string_literal: true
 
 class OfferPolicy < ApplicationPolicy
+  class Scope < Scope
+    def resolve
+      scope.where(status: :published)
+    end
+  end
+
   def index?
     true
   end