+++ db/migrate/20190306111257_add_service_order_target_to_service.rb
@@ -1,0 +1,5 @@
+class AddServiceOrderTargetToService < ActiveRecord::Migration[5.2]
+  def change
+    add_column :services, :order_target, :string, null: false, default: ""
+  end
+end