+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 ## [Unreleased]
 
 ### Added
+- Offer status (@martaswiatkowska)
 
 ### Changed
 
@@ -23,7 +24,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Fixed
 
-- Small view error in my order detail page 
+- Small view error in my order detail page
 
 ## [1.4.0] - 2019-02-26