+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Added
 - Offer status (@martaswiatkowska)
+- markdown rendering in offer's description (@goreck888)
 
 ### Changed