+++ CHANGELOG.md
@@ -19,6 +19,13 @@ Please view this file on the master branch, on stable branches it's out of date.
 
 ### Security
 
+
+## [1.5.1] - 2019-03-06
+
+### Security
+
+- Rails upgraded with fixes for: CVE-2019-5418, CVE-2019-5419 and CVE-2019-5420 (@mkasztelnik)
+
 ## [1.5.0] - 2019-03-06
 
 ### Added